fix: make encoder output GOARCH-reproducible via float32 FMA barriers (#78)
Adds explicit float32() rounding barriers at the encoder's float32 a*b+c
sites so the Go compiler cannot contract them into a fused multiply-add on
arm64 (it does not on default amd64), matching the -ffp-contract=off FFmpeg
reference. The encoder now produces byte-identical output on arm64 and amd64
given identical input; the barriers are a no-op on amd64.

Adds two guards: a build-time scan that fails on any float32 FMA, and a
committed cross-arch ADTS determinism golden that CI runs on both arches.

The originally reported corpus splits were partly a harness artifact: the
tonal signal generator's float64 math is FMA-contracted on arm64, so its
float32 input samples differ across arches independent of the encoder.

Closes #59.

+++ internal/enc/encoder.go
@@ -360,7 +360,7 @@ func (e *Encoder) EncodeFrame(dst []byte, samples [][]float32) ([]byte, error) {
 		if counted > 0 {
 			side := float32(e.lastFramePBCount) - float32(counted)
 			if e.nmr.SideInited {
-				e.nmr.SideEMA += 0.125 * (side - e.nmr.SideEMA)
+				e.nmr.SideEMA += float32(0.125 * (side - e.nmr.SideEMA)) // no FMA
 			} else {
 				e.nmr.SideEMA = side
 				e.nmr.SideInited = true
